pyproject.toml
[project]
name = "neuron"
version = "0.0.1"
description = "A fastapi+pydantic-ai server for neural network inference."
authors = [
    { name = "Roman", email = "iamromandev@gmail.com" },
]
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
    ### core ###
    "toml>=0.10.2",
    "loguru>=0.7.3",
    "fastapi[all]>=0.121.1",
    ### database ###
    "tortoise-orm[accel,aiomysql]>=0.25.1",
    "aerich[mysql,toml]>=0.9.2",
    ### cache ###
    "redis>=7.0.1",
    ### ai ###
    "pydantic-ai[examples]>=1.22.0",
    ### other ###
    "aiohttp>=3.13.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.4",
    "mypy>=1.18.2",
    "ty>=0.0.1a25",
]

[tool.uv]
package = false

[tool.aerich]
tortoise_orm = "src.data.DB_CONFIG"
location = "./src/data/migration"
src_folder = "./."
